Added new dictionary and moved header widgets, also changed its name from UpperBar
