+++ lib/pages/question.dart
@@ -1,51 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:quizly_app/widgets/header.dart';
 
 
-
-class UpperBar extends StatelessWidget {
-  const UpperBar({super.key});
-
-
-  Widget upperBar() {
-    return AppBar(
-      backgroundColor: Colors.cyan,
-      toolbarHeight: 75,
-      shape: const RoundedRectangleBorder(
-        borderRadius: BorderRadius.only(
-          bottomLeft: Radius.circular(35),
-          bottomRight: Radius.circular(35),
-        ),
-      ),
-      title: Row(
-        children: [
-          IconButton(
-            icon: const Image(image: AssetImage('assets/images/back.png'),height: 90,),
-            onPressed: (){}, // null disables the button
-            iconSize: 58  ,
-          ),
-          const Expanded(
-            child: Image(
-              image: AssetImage('assets/images/logo.png'),
-              height: 60,
-            ),
-          ),
-          IconButton(
-            icon: const Image(
-                image: AssetImage('assets/images/settings.png'),
-
-            ),
-            iconSize: 58,
-            onPressed: (){},
-          ),
-        ],
-      ),
-    );
-  }
-  @override
-  Widget build(BuildContext context) {
-    return upperBar();
-  }
-}
 
 class Questionpage extends StatelessWidget {
   const Questionpage({super.key});
@@ -79,7 +35,7 @@ class Questionpage extends StatelessWidget {
       child: Column(
 
         children:  [
-        const  UpperBar(),
+        const  Header(),
           const SizedBox(height: 20),
           const Text(
            "Pytanie 3:",
